Refactor add dog
Refactor the way that a new dog's ID number is generated to avoid scenarios where a dog that previously had an ID equal to the length of the dogs array minus 1 was deleted and replaced. The new replacement dog would have had it's ID set to the array's length, but there would have already been a dog with that ID. The new method randomly generates an ID between 1 and 100 and then checks against existing IDs.